+++ ShapeHeader/Path.h
@@ -19,7 +19,7 @@ namespace MyFigure
     public:
         friend bool isDigit(const char &c);
         friend bool isAlpha(const char &c);
-        friend int extractNumber(const string &data, int &i);
+        friend float extractNumber(const string &data, int &i);
         Path(xml_node<>* rootNode, Gdiplus::Graphics& graphics);
         ~Path() { path.Reset(); };
         void draw() override;
